chore - adicionado o Bigdecimal

+++ src/main/java/com/ifpb/payment/decorator/Payment.java
@@ -1,6 +1,8 @@
 package com.ifpb.payment.decorator;
 
+import java.math.BigDecimal;
+
 public interface Payment {
-    double getAmount();
+    BigDecimal getAmount();
     String getDescription();
 }
