+++ src/main/java/com/ifpb/payment/strategy/PaymentPixStrategy.java
@@ -1,8 +1,10 @@
 package com.ifpb.payment.strategy;
 
+import java.math.BigDecimal;
+
 public class PaymentPixStrategy implements PaymentStrategy {
     @Override
-    public double calculateFinalAmount(double amount) {
+    public BigDecimal calculateFinalAmount(BigDecimal amount) {
         return amount; // não contem taxa
     }
 }
